Made message class as internal actor class to guarantee that it won't be sent from outside. Added processing of events approvals after Recover phase ends. #291
